aula-12/pyproject.toml
[tool.poetry]
name = "aula-12-grpc"
version = "0.1.0"
description = "gRPC Protocol Buffers and Remote Procedure Calls"
authors = ["Backend Course <backend@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.8.1"
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"
protobuf = "^4.25.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
